updated the editor input and editor nodes for the video and audio
